moved base project in oprject_Type root folder
