+++ mapswipe_workers/mapswipe_workers/project_types/arbitrary_geometry/project.py
@@ -7,8 +7,8 @@ from osgeo import ogr
 
 from mapswipe_workers.definitions import DATA_PATH, CustomError, logger
 from mapswipe_workers.project_types.arbitrary_geometry import grouping_functions as g
-from mapswipe_workers.project_types.base.project import BaseGroup, BaseProject
-from mapswipe_workers.project_types.base.tile_server import BaseTileServer
+from mapswipe_workers.project_types.project import BaseGroup, BaseProject
+from mapswipe_workers.project_types.tile_server import BaseTileServer
 
 
 @dataclass
