+++ mapswipe_workers/mapswipe_workers/project_types/tile_map_service/change_detection/project.py
@@ -1,11 +1,11 @@
 from dataclasses import asdict, dataclass
 
 from mapswipe_workers.firebase.firebase import Firebase
-from mapswipe_workers.project_types.base.tile_server import BaseTileServer
 from mapswipe_workers.project_types.tile_map_service.project import (
     TileMapServiceBaseProject,
     TileMapServiceBaseTask,
 )
+from mapswipe_workers.project_types.tile_server import BaseTileServer
 from mapswipe_workers.utils.tile_functions import tile_coords_zoom_and_tileserver_to_url
