+++ mapswipe_workers/mapswipe_workers/project_types/tile_map_service/tutorial.py
@@ -3,12 +3,12 @@ from typing import Dict, List
 
 from mapswipe_workers.definitions import logger
 from mapswipe_workers.firebase.firebase import Firebase
-from mapswipe_workers.project_types.base.tile_server import BaseTileServer
-from mapswipe_workers.project_types.base.tutorial import BaseTutorial
 from mapswipe_workers.project_types.tile_map_service.project import (
     TileMapServiceBaseGroup,
     TileMapServiceBaseTask,
 )
+from mapswipe_workers.project_types.tile_server import BaseTileServer
+from mapswipe_workers.project_types.tutorial import BaseTutorial
 from mapswipe_workers.utils import tile_functions as t
