+++ mapswipe_workers/mapswipe_workers/project_types/tile_server.py
